[refactor] how files are included.
