+++ src/app/engine/ng-engine.module.ts
@@ -12,7 +12,7 @@ import { NgEngineStore } from './ng-engine.store'
 import { NgEngineService } from './ng-engine.service'
 
 // Root Reducers
-import { reducers, metaReducers } from '../store/reducers'
+import { reducers, metaReducers } from '../root/store/reducers'
 
 // Combine Root Reducers with Pack Reducers
 export const REDUCER_TOKEN = new InjectionToken<ActionReducerMap<any>>('Pack Reducers')
