--- src/app/packs/404/404.actions.ts
@@ -1,1 +1,0 @@
-export const ACTIONS = {}
