--- src/app/packs/404/404.effects.ts
@@ -1,1 +1,0 @@
-export const EFFECTS = []
