--- src/app/packs/404/404.reducers.ts
@@ -1,1 +1,0 @@
-export const REDUCERS = {}
