+++ src/app/packs/404/index.ts
@@ -1,9 +1,7 @@
 import { NgPack} from '../../engine/ng-pack'
 import * as CONFIG from './config'
 import * as PKG from './package.json'
-import { ACTIONS } from './404.actions'
-import { EFFECTS } from './404.effects'
-import { REDUCERS } from './404.reducers'
+import { ACTIONS, EFFECTS, REDUCERS } from './store'
 
 export class FourZeroFour extends NgPack {
   constructor(app) {
