+++ src/app/packs/404/store/actions/index.ts
@@ -1,0 +1,1 @@
+export {}
