+++ src/app/packs/404/store/effects/index.ts
@@ -1,0 +1,1 @@
+export {}
