+++ src/app/packs/404/store/index.ts
@@ -1,4 +1,4 @@
-import { actions } from './actions'
+import * as actions from './actions'
 import * as effects from './effects'
 import * as reducers from './reducers'
