+++ src/app/packs/404/store/reducers/index.ts
@@ -1,0 +1,1 @@
+export {}
