--- src/app/packs/500/500.actions.ts
@@ -1,1 +1,0 @@
-export const ACTIONS = {}
