--- src/app/packs/500/500.effects.ts
@@ -1,1 +1,0 @@
-export const EFFECTS = {}
