--- src/app/packs/500/500.reducers.ts
@@ -1,1 +1,0 @@
-export const REDUCERS = {}
