+++ src/app/packs/500/index.ts
@@ -1,9 +1,7 @@
 import { NgPack } from '../../engine/ng-pack'
 import * as CONFIG from './config'
 import * as PKG from './package.json'
-import { ACTIONS } from './500.actions'
-import { EFFECTS } from './500.effects'
-import { REDUCERS } from './500.reducers'
+import { ACTIONS, EFFECTS, REDUCERS } from './store'
 
 export class FiveZeroZero extends NgPack {
   constructor(app) {
